requirements.txt
black
jsonmerge
invoke
python-dotenv
PyYAML